Adds struct & class for painting any actor
Adds struct and class that serve as the base for the architecture to
paint any actor in the game. Any actor that will be painted will have
this component that will interact with any projectile that hits the
actor. Sets up the basis for using queues to carry painting instructions
to the scene captures.