Fix spawning projectiles on server
There were two issues at play hindering spawning projectiles on the
server. One the Projectile blueprint did not have replication marked as
true. Second, I was using a boolean to track whether the gun was still
being fired, but I did not set that boolean to be replicated, therefore
when the server checked if the boolean was true it was always false.
Instead I now pass a boolean into the RPC to determine if projectiles
should still be spawned or not.