Adds logic and variables related to health
Health in the game is not defined by a number that ticks down each time
a projectile shot, but is calculated as an overall percentage of how
much area of the mesh is covered in paint. In order to achieve this
effect variables to set the total number of tiles that need to be
painted for player death have been added. Next step is to figure out
replication, which variables need to be replicated and why.