Add PaintColorManager class to aid setting colors for teams
Adds a new class called PaintColorManager that is responsible for setting the right colors depending on teams for all players. This changes the setup so all colors now live in the player character class as a property that are referenced whenever a paint event may occur. This is a helpful change because all colors live in one place when the game begins.